requirements.txt
fastapi
uvicorn
aiohttp
beautifulsoup4
pycryptodome
pydantic